refactoring and adding adam
